# docs/v4/openapi.yaml
openapi: 3.1.0
info:
  title: AgentMesh Public API
  version: 4.0.0-dev
  description: Project-scoped public API introduced by the V4 Platform Ecosystem sprint.
servers:
  - url: https://agentmesh.example.com
security:
  - ServiceAccountBearer: []
paths:
  /openapi/v1/tasks/run:
    post:
      summary: Run a task in the Service Account project
      operationId: runTask
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunTaskRequest'
      responses:
        '200':
          description: Task execution result or idempotent replay
          headers:
            X-Idempotent-Replay:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '400': {$ref: '#/components/responses/ErrorResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '404': {$ref: '#/components/responses/ErrorResponse'}
        '409': {$ref: '#/components/responses/ErrorResponse'}
...
components:
  securitySchemes:
    ServiceAccountBearer:
      type: http
      scheme: bearer
      bearerFormat: AgentMesh Service Account API Key
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, maxLength: 160}
  schemas:
    RunTaskRequest:
      type: object
      required: [task]
      properties:
        task: {type: string, minLength: 1}
        conversationId: {type: integer, format: int64, minimum: 1}
        scheduler: {type: string, default: adaptive}
        planner: {type: string, default: multi_objective}
        executionMode: {type: string, default: auto}
        synthesisMode: {type: string, default: auto}
        constraints: {type: object, additionalProperties: true}
    ObjectEnvelope:
      type: object
      required: [code, data]
      properties:
        code: {type: integer, const: 0}
        message: {type: string}
        data: {type: object, additionalProperties: true}
...
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code: {type: integer}
        message: {type: string}
        data: {}
  responses:
    ErrorResponse:
      description: AgentMesh API error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}